Use setInterval instead of route + cron
