+++ lib/application.js
@@ -15,6 +15,7 @@ const database = require('./database');
 const { webUiServer } = require('./server');
 const { gRPCServer } = require('./server/index.js');
 const { GRPCConfig } = require('./config');
+const { handleLostRunsAndEnvironments } = require('./server/services/housekeeping/handleLostRunsAndEnvironments.js');
 
 /**
  * Bookkeeping Application
@@ -54,6 +55,16 @@ class BookkeepingApplication {
             if (gRPCOrigin) {
                 await this.gRPCServer.listen(gRPCOrigin);
             }
+
+            // Start background housekeeping
+            setInterval(async () => {
+                try {
+                    await handleLostRunsAndEnvironments();
+                    this.logger.debug('Updated lost runs and environments');
+                } catch (error) {
+                    this.logger.error(`Error while handling lost logs: ${error}`);
+                }
+            }, 15 * 60 * 1000);
         } catch (error) {
             this.logger.error(`Error while starting: ${error}`);
             return this.stop();
