+++ lib/server/controllers/environments.controller.js
@@ -30,8 +30,8 @@ const { dtoValidator } = require('../utilities');
 const { ApiConfig } = require('../../config/index.js');
 const { environmentService } = require('../services/environment/EnvironmentService.js');
 const { updateExpressResponseFromNativeError } = require('../express/updateExpressResponseFromNativeError.js');
-const { purgeEnvironments } = require('../services/environment/purgeEnvironments.js');
-const { purgeRuns } = require('../services/run/purgeRuns.js');
+const { transitionToErrorLostEnvironments } = require('../services/environment/transitionToErrorLostEnvironments.js');
+const { setO2StopOfLostRuns } = require('../services/run/setO2StopOfLostRuns.js');
 const { buildUrl } = require('../../utilities/buildUrl.js');
 const { services } = require('../../config/services.js');
 
@@ -145,55 +145,10 @@ const getEnvironmentById = async (request, response) => {
     }
 };
 
-/**
- * Purge the lost environments and runs
- *
- * For all environments that has not been destroyed and have not gone to error, check through AliECS GUI if they are still alive, and if not,
- * mark them as gone to error For all runs that do not have timeO2Stop or timeTrgStop, check through AliECS Gui if they are still running, and
- * if not, mark them as stopped NOW
- *
- * @param {object} request the express request
- * @param {object} response the express response
- * @return {Promise<void>} resolve once the environments and runs has been purged
- * @deprecated
- */
-const purgeLostEnvironmentsAndRunsHandler = async (request, response) => {
-    try {
-        // TODO remove with node 18
-        const { default: fetch } = await import('node-fetch');
-        const existingEnvironmentsResponse = await fetch(buildUrl(
-            `${services.aliEcsGUI.url}/api/core/environments`,
-            { token: services.aliEcsGUI.token },
-        ));
-
-        if (existingEnvironmentsResponse.ok) {
-            const { environments } = await existingEnvironmentsResponse.json();
-            const environmentIdsToKeep = [];
-            const runNumbersToKeep = [];
-            for (const environment of environments) {
-                const { id, currentRunNumber } = environment;
-                environmentIdsToKeep.push(id);
-                if (currentRunNumber) {
-                    runNumbersToKeep.push(currentRunNumber);
-                }
-            }
-
-            // Maximum amount of minutes after run start/environment creation to purge them
-            const modificationTime = 2880;
-
-            await purgeEnvironments(environmentIdsToKeep, modificationTime);
-            await purgeRuns(runNumbersToKeep, modificationTime);
-            response.status(200).send();
-        }
-    } catch (error) {
-        updateExpressResponseFromNativeError(response, error);
-    }
-};
 
 module.exports = {
     createEnvironment,
     updateEnvironment,
     getAllEnvironments,
     getEnvironmentById,
-    purgeLostEnvironmentsAndRunsHandler,
 };
