+++ lib/server/routers/environments.router.js
@@ -36,11 +36,5 @@ module.exports = {
                 },
             ],
         },
-        {
-            path: '/purge',
-            method: 'post',
-            controller: EnvironmentsController.purgeLostEnvironmentsAndRunsHandler,
-            args: { public: true },
-        },
     ],
 };
