+++ lib/server/services/environment/transitionToErrorLostEnvironments.js
@@ -16,13 +16,13 @@ const { sequelize } = require('../../../database/index.js');
 /**
  * Consider all the environments that are currently active but not in the given list of ids to be lost, and transition them to error
  *
- * @param {number[]} environmentIdsToKeep the list of ids of environments to keep
+ * @param {string[]} activeEnvironmentIds the ids of active environments (those environments will not be transitioned to error)
  * @param {number} modificationTimeWindow the time we have (in MINUTES), after the creation of an environment, to purge them. If an environment
  *     has been created more than this amount of minutes before NOW, it is NOT purged
- * @return {Promise<void>} resolve once the purge has been done
+ * @return {Promise<void>} resolve once all inactive environments has been transitioned to error
  * @deprecated
  */
-exports.purgeEnvironments = async (environmentIdsToKeep, modificationTimeWindow) => sequelize.query(`
+exports.transitionToErrorLostEnvironments = async (activeEnvironmentIds, modificationTimeWindow) => sequelize.query(`
     INSERT INTO environments_history_items (environment_id, status, status_message)
     SELECT e.id, 'ERROR', 'Shutdown not received, environment not existing anymore'
     FROM environments_history_items ehi
@@ -31,5 +31,5 @@ exports.purgeEnvironments = async (environmentIdsToKeep, modificationTimeWindow)
     HAVING e.updated_at > date_sub(now(), INTERVAL ${modificationTimeWindow} MINUTE)
        AND group_concat(ehi.status) NOT LIKE '%ERROR%'
        AND group_concat(ehi.status) NOT LIKE '%DESTROYED%'
-       AND e.id NOT IN (${environmentIdsToKeep.map((id) => `'${id}'`).join(',')})
+       AND e.id NOT IN (${activeEnvironmentIds.map((id) => `'${id}'`).join(',')})
 `);
