+++ lib/server/services/housekeeping/handleLostRunsAndEnvironments.js
@@ -1,0 +1,44 @@
+const { services } = require('../../../config/services.js');
+const { buildUrl } = require('../../../utilities/buildUrl.js');
+const { transitionToErrorLostEnvironments } = require('../environment/transitionToErrorLostEnvironments.js');
+const { setO2StopOfLostRuns } = require('../run/setO2StopOfLostRuns.js');
+
+/**
+ * Handle lost environments and runs
+ *
+ * For all environments that has not been destroyed and have not gone to error, check through AliECS GUI if they are still active, and if not,
+ * mark them as gone to error. For all runs that do not have timeO2Stop or timeTrgStop, check through AliECS GUI if they are still running, and
+ * if not, mark them as stopped NOW
+ *
+ * @return {Promise<void>} resolve once the environments and runs has been cleaned
+ * @deprecated
+ */
+exports.handleLostRunsAndEnvironments = async () => {
+    // TODO remove with node 18
+    const { default: fetch } = await import('node-fetch');
+    const existingEnvironmentsResponse = await fetch(buildUrl(
+        `${services.aliEcsGUI.url}/api/core/environments`,
+        { token: services.aliEcsGUI.token },
+    ));
+
+    if (existingEnvironmentsResponse.ok) {
+        const { environments } = await existingEnvironmentsResponse.json();
+        const environmentIdsToKeep = [];
+        const runNumbersToKeep = [];
+        for (const environment of environments) {
+            const { id, currentRunNumber } = environment;
+            environmentIdsToKeep.push(id);
+            if (currentRunNumber) {
+                runNumbersToKeep.push(currentRunNumber);
+            }
+        }
+
+        // Maximum amount of minutes after run start/environment creation to purge them
+        const modificationTime = 2880;
+
+        await transitionToErrorLostEnvironments(environmentIdsToKeep, modificationTime);
+        await setO2StopOfLostRuns(runNumbersToKeep, modificationTime);
+    } else {
+        throw new Error(`Received ${existingEnvironmentsResponse.status} from AliECS GUI`);
+    }
+};
