+++ lib/server/services/run/setO2StopOfLostRuns.js
@@ -16,18 +16,18 @@ const { sequelize } = require('../../../database/index.js');
 /**
  * Consider all the runs that are still running but not in the given list of run numbers to be lost, and set their stop time to now
  *
- * @param {number[]} runNumbersToKeep the list of run numbers corresponding to runs that are actually running
+ * @param {number[]} runNumbersOfRunningRuns the list of run numbers corresponding to runs that are actually running
  * @param {number} modificationTimeWindow the time we have (in MINUTES), after the start of a run, to purge them. If a run started more than this
  *     amount of minutes before NOW, it is NOT purged
- * @return {Promise<void>} resolve once the purge has been done
+ * @return {Promise<void>} resolve once the lost runs has been marked as ended
  * @deprecated
  */
-exports.purgeRuns = (runNumbersToKeep, modificationTimeWindow) => sequelize.query(`
+exports.setO2StopOfLostRuns = (runNumbersOfRunningRuns, modificationTimeWindow) => sequelize.query(`
     UPDATE runs
     SET time_o2_end = now()
     WHERE time_o2_end IS NULL
       AND time_trg_end IS NULL
       AND COALESCE(time_trg_start, time_o2_start) IS NOT NULL
       AND COALESCE(time_trg_start, time_o2_start) > date_sub(now(), INTERVAL ${modificationTimeWindow} MINUTE)
-      AND run_number NOT IN (${runNumbersToKeep.join(',')});
+      AND run_number NOT IN (${runNumbersOfRunningRuns.join(',')});
 `);
